exclude command refactoring from PR
